+++ packages/git/src/getCommitMessages.ts
@@ -1,6 +1,6 @@
 import type { CommitMessage, MonodeployConfiguration, YarnContext } from '@monodeploy/types'
 
-import { gitLog } from './gitCommands'
+import { gitLastTaggedCommit, gitLog } from './gitCommands'
 
 const DELIMITER = '-----------------monodeploy-----------------'
 
@@ -8,8 +8,22 @@ export const getCommitMessages = async (
     config: MonodeployConfiguration,
     context?: YarnContext,
 ): Promise<CommitMessage[]> => {
-    const from = config.git.baseBranch
     const to = config.git.commitSha
+    let from = config.git.baseBranch
+    if (!from) {
+        const { sha, tag } = await gitLastTaggedCommit({
+            cwd: config.cwd,
+            context,
+            prerelease: config.prerelease,
+        })
+        from = sha
+
+        if (to === from && tag !== null) {
+            // the latest commit is the tagged commit, this run of monodeploy should be a no-op
+            return []
+        }
+    }
+
     const logOutput = await gitLog(from, to, {
         cwd: config.cwd,
         DELIMITER,